02-agents/local-agent/requirements.txt
python-dotenv
agno[all]
xpander-sdk[agno]
ollama
openai